move lights to scene's graphical setup and use timer to animate lights
