+++ src/scene/light/directional_light.h
