+++ src/scene/light/point_light.h
