+++ src/scene/light/spot_light.h
