+++ src/scene/main_scene.cpp
@@ -4,7 +4,7 @@
 #include "game_objects/test_coin.h"
 #include "events.h"
 #include "util/color.h"
-#include "pulse_point_light.h"
+#include "light/pulse_point_light.h"
 
 MainScene::MainScene() : Scene() {
 
@@ -27,7 +27,13 @@ void MainScene::client_update() {
 }
 
 void MainScene::setup_scene() {
+    //Setting up map
+    grid = new Grid("assets/maps/dungeon_test.txt");
+    objs.push_back(grid);
 
+}
+
+void MainScene::graphical_setup() {
     // Setup lights.
     DirectionalLight *sun = new DirectionalLight(glm::vec3(0.f, -1.f, -1.f),
                                                  Color::WHITE, 0.5f);
@@ -40,22 +46,14 @@ void MainScene::setup_scene() {
     add_light(bulb2);
 
     SpotLight *spot_light = new SpotLight(glm::vec3(10.f, 4.f, 5.f),
-                                     glm::vec3(0.f, -1.f, 0.f),
-                                     Color::MAGENTA);
+                                          glm::vec3(0.f, -1.f, 0.f),
+                                          Color::MAGENTA);
     add_light(spot_light);
 
     // Setup light debug callback.
     events::debug::toggle_light_rotation_event.connect([&](void) {
         lights_debug_on = !lights_debug_on;
     });
-
-    //Setting up map
-    grid = new Grid("assets/maps/dungeon_test.txt");
-    objs.push_back(grid);
-
-}
-
-void MainScene::graphical_setup() {
 
     for (GameObject *obj : objs) {
         obj->setup_model();
